requirements.txt
# Notebooks & Utils
jupyterlab
notebook
openai
huggingface_hub
python-dotenv

# Agents
crewai
langgraph
pydantic-ai
nest_asyncio
semantic-kernel
google-generativeai

# API & Server
fastapi
uvicorn
